Assistant/requirements.txt
mlflow>=2.21.2
streamlit==1.44.1
openai
databricks-sql-connector
databricks-sdk
toml>=0.10.2
cryptography
tenacity>=8.2.0
typing-extensions>=4.5.0
structlog>=24.1.0
streamlit-javascript>=0.1.5
PyPDF2>=3.0.0
Pillow>=10.0.0
psutil>=5.9.0